use redefined RBF constants

+++ app/src/main/java/com/samourai/wallet/send/SendFactory.java
@@ -259,7 +259,7 @@ public class SendFactory	{
 
             MyTransactionInput input = new MyTransactionInput(SamouraiWallet.getInstance().getCurrentNetworkParams(), null, new byte[0], outPoint, outPoint.getTxHash().toString(), outPoint.getTxOutputN());
             if(PrefsUtil.getInstance(context).getValue(PrefsUtil.RBF_OPT_IN, false) == true)    {
-                input.setSequenceNumber(SamouraiWallet.RBF_SEQUENCE_VAL);
+                input.setSequenceNumber(SamouraiWallet.RBF_SEQUENCE_VAL.longValue());
             }
             inputs.add(input);
         }
